+++ zrml/swaps/fuzz/helper_functions.rs
@@ -1,8 +1,10 @@
 use zeitgeist_primitives::types::{Asset, ScalarPosition, SerdeWrapper};
 
-pub fn asset(seed: (u128, u16)) -> Asset<u128> {
+pub static _CREATE_POOL_FAILURE: &str = "Pool creation failed unexpectedly. Error:";
+
+pub fn construct_asset(seed: (u128, u16)) -> Asset<u128> {
     let (seed0, seed1) = seed;
-    let module = seed0 % 4;
+    let module = seed0 % 5;
     match module {
         0 => Asset::CategoricalOutcome(seed0, seed1),
         1 => {
