+++ zrml/swaps/fuzz/pool_exit_with_exact_pool_amount.rs
@@ -1,0 +1,58 @@
+#![no_main]
+
+use libfuzzer_sys::fuzz_target;
+use zrml_swaps::mock::{ExtBuilder, Origin, Swaps};
+
+mod data_structs;
+use data_structs::ExactAmountData;
+mod helper_functions;
+use helper_functions::{construct_asset, _CREATE_POOL_FAILURE};
+use orml_traits::MultiCurrency;
+use zeitgeist_primitives::{
+    constants::MinLiquidity,
+    traits::Swaps as SwapsTrait,
+    types::{Asset, ScoringRule, SerdeWrapper},
+};
+use zrml_swaps::mock::Shares;
+
+fuzz_target!(|data: ExactAmountData| {
+    let mut ext = ExtBuilder::default().build();
+    let _ = ext.execute_with(|| {
+        // ensure that the account origin has a sufficient balance
+        // use orml_traits::MultiCurrency; required for this
+        for a in &data.pool_creation.assets {
+            let _ = Shares::deposit(
+                construct_asset(*a),
+                &data.pool_creation.origin,
+                MinLiquidity::get(),
+            );
+        }
+        let _ = Shares::deposit(
+            Asset::PoolShare(SerdeWrapper(0)),
+            &data.pool_creation.origin,
+            data.pool_amount,
+        );
+
+        match Swaps::create_pool(
+            data.pool_creation.origin,
+            data.pool_creation.assets.into_iter().map(construct_asset).collect(),
+            construct_asset(data.pool_creation.base_asset),
+            data.pool_creation.market_id,
+            ScoringRule::CPMM,
+            Some(data.pool_creation.swap_fee),
+            Some(data.pool_creation.weights),
+        ) {
+            Ok(pool_id) => {
+                let _ = Swaps::pool_exit_with_exact_pool_amount(
+                    Origin::signed(data.origin),
+                    pool_id,
+                    construct_asset(data.asset),
+                    data.pool_amount,
+                    data.asset_amount,
+                );
+            }
+            Err(e) => panic!("{_CREATE_POOL_FAILURE} {:?}", e),
+        }
+    });
+    let _ = ext.commit_all();
+});
